+++ client/src/components/menu/menu.js
@@ -6,7 +6,9 @@ import { bindActionCreators } from 'redux';
 import { Nav, NavItem, Navbar } from 'react-bootstrap';
 import PropTypes from 'prop-types';
 
-import { setGuest } from '../actions/authentication'; // sets fake guest account
+import { setGuest } from '../../actions/authentication'; // sets fake guest account
+import Cover from './cover';
+import './menu.scss';
 
 class Menu extends Component {
 
@@ -51,50 +53,7 @@ class Menu extends Component {
         )
       }
       // display only login and welcome card if null username
-      return (
-        <div id="cover">
-          <div id="logincard">
-            <div>
-              <div id="logo"><i className="fa fa-pinterest" aria-hidden="true" /></div>
-              <div id="welcome">Welcome to Pinterest</div>
-              <div id="subheader">Find new ideas to try</div>
-              <div id="disclaimer">Cloned For Educational Purposes Only!</div>
-              <div id="gitsource">
-                <a href="https://github.com/Dereje1/Pinterest-Clone" target="_blank" rel="noopener noreferrer">
-                  <i className="fa fa-github" aria-hidden="true" />
-                  {' Github'}
-                </a>
-              </div>
-            </div>
-            <div>
-              <button
-                type="submit"
-                id="guestbutton"
-                onClick={this.handleGuest}
-              >
-                <span id="guest">
-                  <i className="fa fa-question-circle" aria-hidden="true" />
-                </span>
-                <span className="buttontext">
-                  {'Continue As Guest'}
-                </span>
-              </button>
-              <button
-                type="submit"
-                id="loginbutton"
-                onClick={this.handleLogin}
-              >
-                <span id="twitter">
-                  <i className="fa fa-twitter" aria-hidden="true" />
-                </span>
-                <span className="buttontext">
-                  {'Continue With Twitter'}
-                </span>
-              </button>
-            </div>
-          </div>
-        </div>
-      );
+      return <Cover handleGuest={this.handleGuest} handleLogin={this.handleLogin} />;
     }
     // for twitter authenticated users
     return (
