+++ client/src/index.js
@@ -16,8 +16,8 @@ import {
 // Import all Created react components that are associated with the router
 // standard menu (non-authenticated) components
 import Main from './main';
-import Home from './components/home';
-import Mypins from './components/mypins';
+import Home from './components/home/home';
+import Mypins from './components/mypins/mypins';
 // import combined reducer to pass to store here
 import reducers from './reducers/index';
